Create Context structures
These can be used during the pre-analysis stage.
